Add notification (toastr)

+++ AdminLTE.MVC/Controllers/LocalCommunityController.cs
@@ -3,6 +3,7 @@ using AdminLTE.Models;
 using AdminLTE.Models.ViewModels;
 using AdminLTE.MVC.Data;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -11,6 +12,15 @@ using System.Threading.Tasks;
 
 namespace AdminLTE.Controllers
 {
+    public enum MessageType
+    {
+        NONE,
+        ADD,
+        REMOVE,
+        EDIT,
+        EXIST_WORKERS,
+        ERROR,
+    }
     public class LocalCommunityController : Controller
     {
         private readonly ApplicationDbContext _db;
@@ -19,7 +29,7 @@ namespace AdminLTE.Controllers
             _db = db;
         }
 
-        public IActionResult Index(int page = 1)
+        public IActionResult Index(int page = 1, MessageType message = MessageType.NONE)
         {
             int pageSize = 5;
 
@@ -34,6 +44,25 @@ namespace AdminLTE.Controllers
                 PageInfo = new PageInfo { PageNumber = page, PageSize = pageSize, TotalItems = lcCount }
             };
 
+            switch (message)
+            {
+                case MessageType.ADD:
+                    ViewBag.JavaScriptFunction = string.Format("ShowMessage('{0}', '{1}');", "Громада успішно добавлена", "success");
+                    break;
+                case MessageType.REMOVE:
+                    ViewBag.JavaScriptFunction = string.Format("ShowMessage('{0}', '{1}');", "Громада успішно видалена", "success");
+                    break;
+                case MessageType.EDIT:
+                    ViewBag.JavaScriptFunction = string.Format("ShowMessage('{0}', '{1}');", "Громада успішно відредагована", "success");
+                    break;
+                case MessageType.EXIST_WORKERS:
+                    ViewBag.JavaScriptFunction = string.Format("ShowMessage('{0}', '{1}');", "Неможливо видалити громаду, видаліть працівників", "error");
+                    break;
+                case MessageType.ERROR:
+                    ViewBag.JavaScriptFunction = string.Format("ShowMessage('{0}', '{1}');", "Виникла помилка", "error");
+                    break;
+            }
+
             return View(lcvm);
         }
         public IActionResult Add(string title)
@@ -43,35 +72,52 @@ namespace AdminLTE.Controllers
                 LocalCommunity lc = new LocalCommunity() { Title = title };
                 _db.LocalCommunities.Add(lc);
                 _db.SaveChanges();
+                return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.ADD });
             }
 
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.ERROR });
         }
 
         //TODO: Notifications (до прикладу при добавленні і т.д.)
-        //TODO: Create pagination
         public IActionResult Edit(int id, string title)
         {
-            LocalCommunity lc = _db.LocalCommunities.Find(id);
-            if (lc == null)
-                return NotFound();
+            try
+            {
+                LocalCommunity lc = _db.LocalCommunities.Find(id);
+                if (lc == null)
+                    return NotFound();
 
-            lc.Title = title;
-            _db.SaveChanges();
-
-            return RedirectToAction(nameof(Index));
+                lc.Title = title;
+                _db.SaveChanges();
+                return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.EDIT });
+            }
+            catch (Exception ex)
+            {
+                return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.ERROR });
+            }
         }
         [HttpPost]
         public IActionResult Remove(int id)
         {
-            LocalCommunity lc = _db.LocalCommunities.Find(id);
-            if (lc == null)
-                return NotFound();
+            try
+            {
+                LocalCommunity lc = _db.LocalCommunities.Include(x => x.Employees).SingleOrDefault(x => x.Id == id);
+                if (lc == null)
+                    return NotFound();
 
-            _db.LocalCommunities.Remove(lc);
-            _db.SaveChanges();
+                if (lc.Employees.Count != 0)
+                    return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.EXIST_WORKERS });
 
-            return RedirectToAction(nameof(Index));
+                _db.LocalCommunities.Remove(lc);
+                _db.SaveChanges();
+
+                return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.REMOVE });
+            }
+            catch (Exception)
+            {
+                return RedirectToAction(nameof(Index), new { page = 1, message = MessageType.ERROR });
+            }
+
         }
         [HttpPost]
         public IActionResult ChooseOption(string title, int? id)
